fix dependency cycle and forum-gate perf metric code
